Applied patch: [1629038] Search classpath for eXist's home Submitted By: Jon Fon
"This patch changes the Configuration class to search the classpath for the first occurrence of conf.xml. Having found that path to conf.xml, the Configuration class sets eXist's home to the path's parent directory. This would be a last ditch effort (i.e., the last condition of getExistHome(String)).

I made this patch because the current search of the classpath in Configuration(String, String) requires specifying a configuration filename, and while the DatabaseImpl class supports setting this as a property, the Ant tasks do not appear to support such. Finally, just searching the classpath for a default configuration filename seems like the lazy way to do things, to which I am partial."

svn path=/trunk/eXist/; revision=5154