[bugfix] wrong computation of sibling node ids causes database corruptions after inserting nodes via XUpdate or XQuery update extensions. This is a serious issue, which is not easy to reproduce as you need a certain sequence of node removals followed by node insertions. However, it is likely that users experience the bug sooner or later.
The problem appears if one is inserting nodes into a gap in the node tree, say 1.5.6 is the first child node of 1.5 (1.5.1 to 1.5.5 were removed previously). eXist should fill up the available node ids, i.e. 1.5.1 to 1.5.5, but it should not use 1.5.6 (since that node does already exist). Instead the node between 1.5.5 and 1.5.6 should become 1.5.5/1. Otherwise, a duplicate 1.5.6 would be written and cause the db to become corrupt.

svn path=/trunk/eXist/; revision=7547