[performance] new NodeSet implementation, which is more memory efficient than the old ExtArrayNodeSet. ExtArrayNodeSet wrapped an object around the nodes for a given document and stored them in an array. This allowed direct access to a document and thus an efficient implementation of node set comparisons. However, memory consumption was very high, especially if a database contained lots of very small documents.
The new NodeSet implementation, NewArrayNodeSet, preserves the advantages of ExtArrayNodeSet in that it still allows direct access to documents, but at the same time consumes much less memory per document. All nodes are stored in a single ordered array. But similar to a skip list, we also keep an array of document offsets which point into the node array and another array to keep track of the number of nodes per document. This allows for direct access to the nodes of a document through a binary search on the documents array. 

Port of rev 7890 and others from the eXist-stable-1.2 branch.

svn path=/trunk/eXist/; revision=7922