Fix so that admin browsing will correctly display a resource with a URLEncoded name.
Details:  The admin browsing xquery will URLEncode all resource names that it is displaying in a clickable list. Problem is, if you have an already URLEncoded name (eg. a name with embedded spaces, say "my%20resource.xml", then the amin app will encode it a 2nd time, and the HTML will have  "my%2520resource.xml" in it.  When you click on this, the RESTServer cannot locate the resource due to the double encoding.

The solution was to add a URLDecode of the path in the RESTServer, which fixes things up so that clicking on such a resource name will locate it correctly.

svn path=/trunk/eXist/; revision=8098