[bugfix] When XML-RPC interface is used to upload and store a file using 'upload' and 'parseLocal' XML-RPC calls, server makes its own assumptions about resource kind to be stored, based on MIME type. So it can happens that a client sends an XML resource, but as server does not recognize its MIME type, it is stored as binary, or vice-versa: a client marks HTML as binary, but servers do parse it as XML.
As the main limitation is the XML-RPC interface parseLocal, which does not
allow setting the resource kind, this bugfix adds on one hand a new XML-RPC
interface called parseLocalExt, which takes an additional parameter, the way
to treat the resource (either XML or binary). Also, RemoteCollection has been
taught about this new XML-RPC method, but taking into account compatibility.

svn path=/trunk/eXist/; revision=9150