[bugfix] enclosed expressions must check namespaces for conflicts. If conflict found use 'XXX'{number} prefix.
3.7.4 In-scope Namespaces of a Constructed Element
...
For each namespace used in the name of the constructed element or in the names of its attributes, a namespace binding must exist. If a namespace binding does not already exist for one of these namespaces, a new namespace binding is created for it. If the name of the node includes a prefix, that prefix is used in the namespace binding; if the name has no prefix, then a binding is created for the empty prefix. If this would result in a conflict, because it would require two different bindings of the same prefix, then the prefix used in the node name is changed to an arbitrary implementation-dependent prefix that does not cause such a conflict, and a namespace binding is created for this new prefix.

svn path=/trunk/eXist/; revision=12170