[bugfix] eXist XMLRPC interface was firing next exceptions on server side when describeResource remote method was called:
javax.servlet.ServletException: org.apache.xmlrpc.XmlRpcException: Failed to write XML-RPC response: Long values aren't supported, if isEnabledForExtensions() == false

The bug is in the code since r13229, when res64-eXist branch was merged.
Answered hash from describeResource was holding a key (content-length-64bit)
with an associated Long value, which is not allowed when XMLRPC extensions
are not enabled. This fix changes both client and server sides, so the
expected/sent value is an String.

As a side effect, as shutdown remote method with timeout potentially suffered
the same problem (timeout value was a Long), it has been improved/fixed.


svn path=/trunk/eXist/; revision=14202