Fix bugs in CropFunction
CropFunction.java has two errors:

1. It is presented as a x1,y1,x2,y2 set of bounding boxes, but java.awt.image.CropImageFilter has this signature:
public CropImageFilter(int x,
               int y,
               int w,
               int h)
Constructs a CropImageFilter that extracts the absolute rectangular region of pixels from its source Image as specified by the x, y, w, and h parameters.
Parameters:
x - the x location of the top of the rectangle to be extracted
y - the y location of the top of the rectangle to be extracted
w - the width of the rectangle to be extracted
h - the height of the rectangle to be extracted
https://docs.oracle.com/javase/7/docs/api/java/awt/image/CropImageFilter.html

Thus, the last two parameters were misaligned. This commit converts the x2,y2 given through the xquery function into height and width values, and checks that these values are not < 1.

2. It generated the new cropped image with the width and height values inverted, so that any non-square cropping created a black band and was also misaligned. This commit fixes the dimensions of the new cropped image.