[bugfix] fn:transform - correct URI error codes
Some resolver cases manage to pick up a base URI for stylesheet-text when they shouldn’t make that assumption. They then fail to open a non-existent (wrongly generated) filename, and finally return the wrong error.

Now, we check for this case in the URI resolver and force the correct error immediately.