[bugfix] fn:transform - cache key was too liberal.
Some of the XQTS tests were getting failures due to falsely matching stylesheets in the stylesheet cache.

This change properly distinguishes stylesheets.

Where the stylesheet is a document, its base uri is the important part of the cache key. Where the stylesheet is a piece of text, we checksum the text.

 If the stylesheet doesn’t have a resolved stylesheet base URI or text to checksum, effectively we don’t cache them (well we do, but we use a unique date now()) as an element of the cache, so they will never match another request.