[feature] fn:transform post-process function
Read the post-process parameter, and if it is supplied and is a function, apply it to the output of the transformation as a final step in generating the output of fn:transform.

The post-process function is applied possibly multiple times, once to the main output and once to every secondary output document. The final result is a map with the same keys as before post processing, but with post processed transformed values replacing transformed values.