[bugfix] - namespace binding conflict errors
@see https://github.com/eXist-db/exist/issues/1482

Report an error (XUDY0023) when an insertion is attempted with a node or attribute name in a namespace that conflicts with one which exists in the document.

@see https://www.w3.org/TR/xquery-update-30/#dt-conflict

Check node/attributes inserted recursively. This change does not look at namespaces introduced in the subtrees being inserted, so there is still potential for a conflict in the final tree.
We will consider rejecting conflicting namespaces introduced in the insertion in a later change.