[optimise] Fix BTree cache CPU trashing eviction
Cache refuses to evict inner nodes, and scans the entire cache looking for non-inner nodes, which become increasingly rare.

Ditch the inner node policy, instead just ensure LRU is updated on get(). Why was it not already doing so ? That is the point of an LRU.

I was seeing queries where > 70% of the CPU was spend on BTreeCache.removeNext(). That’s completely fixed; such queries are now 5x faster.