[bugfix] Leaky trigger state when XQuery method not found
XQuery trigger not found is acceptable behaviour (logs a debug message only, could get very chatty logs otherwise).
BUT the exit path from the exception fails to tidy up the per-thread trigger state. This has been observed to leak memory; the per-thread doesn’t get cleaned up until the thread is deleted, and the trigger cyclic-check stack just builds up.

So we clean up the trigger state when a trigger is not found.

Closes https://github.com/eXist-db/exist/issues/5459