check for sorted node set before sorting.
This is often an optimisation, because many node sets are ordered on construction. We could go one better and maintain `isSorted` as we build the set, but this seems to win a lot of performance for the workload we are focused on, to wit looking for preceding and succeeding siblings.