[WIP] ranges with context fix up (context issues)
DocumentNodeRange for multiple overlapping ranges generates multiple instance of NodeProxy for the same node-id, each with their own specific context.

We are also (in fact) doing multiple scans, which might in fact be multiple large scans, to generate these results.

IF we could merge the ranges, we would still need to manage the multiple overlapping NodeSelectors - we have one for each range/context item.

I think it is better to revert all this, and go back to the simpler ranges which still require patching up the result set with the context.